Expose top-level exprs to in-browser analysis.
This is done by adding the top-level expr IIDs to
jalangi_sourcemap.js.  This information is rather
small compared to the rest of the source map, so it
shouldn't affect load times much.  We add the info
to this file since there is already a mechanism in
place for loading the sourcemap into the browser.